requirements.txt
Cython==0.23
Pygame==1.9.1
buildozer==0.31
colorama==0.3.6
jsonpickle==0.9.2
Kivy==1.9.1
Kivy-Garden==0.1.4
qrcode==5.2.2
requests==2.9.1
six==1.10.0
Twisted==15.5.0
txWS==0.9.1
websocket-client==0.35.0
wheel==0.29.0
ws4py==0.3.4
zope.interface==4.1.3
